+++ extensions/src/extensions/getSearch.ts
@@ -1,0 +1,14 @@
+import {SelectFunction, SelectPluginType} from '@anglr/select';
+
+/**
+ * Gets current search value of Select (LiveSearch plugin), its reactive
+ */
+export function getSearch(): SelectFunction<string>
+{
+    return select =>
+    {
+        const liveSearch = select.getPlugin(SelectPluginType.LiveSearch);
+
+        return liveSearch.search();
+    };
+}