+++ extensions/src/index.ts
@@ -1,2 +1,4 @@
+export * from './extensions/getSearch';
 export * from './extensions/getValue';
+export * from './extensions/setReadonly';
 export * from './extensions/setValue';